requirements.txt
openai
supabase-python